requirements.txt
pandas==2.2.3
zabbix-utils==2.0.1python-dotenv==1.0.1
python-dotenv==1.0.1
